services/processor/requirements.txt
# DeepStream Python bindings
gst-python==1.22.0
pyds==1.1.6

# Message broker
pulsar-client==3.4.0
aiokafka==0.10.0

# Async database
sqlalchemy==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0

# Configuration and utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Image processing (for snapshots)
pillow==10.1.0
opencv-python==4.8.1.78

# NumPy (specific version for DeepStream compatibility)
numpy==1.26.0